Comprehensive list of fixes and additions to the MP
- New command /hideme
- Blip is no longer being rendered for the local player and the car he is in
- Fix Oakwood profile saving (remove old OakwoodProfile.data from game directory)

+++ Code/Shared/Oakwood/Framework.hpp
@@ -57,7 +57,7 @@ public:
     }
 
     inline T *GetObjectByID(int idx) {
-        if (idx < 0 || idx >= objects.size()) {
+        if (idx < 0 || idx >= (int)objects.size()) {
             return nullptr;
         }
